Downloads/Role_based_book_api/Book_api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
python-jose
